+++ examples/helloworld/v2/nats/user/user.gen.go
