+++ examples/ping/v2/asyncapi.yaml
