+++ examples/ping/v2/kafka/app/app.gen.go
