+++ examples/ping/v2/kafka/user/user.gen.go
