+++ examples/ping/v2/nats-jetstream/app/app.gen.go
