+++ examples/ping/v2/nats-jetstream/user/user.gen.go
