+++ examples/ping/v2/nats/app/app.gen.go
