+++ examples/ping/v2/nats/user/main.go
@@ -1,11 +1,11 @@
-//go:generate go run ../../../../cmd/asyncapi-codegen -g user,types -p main -i ../../asyncapi.yaml -o ./user.gen.go
+//go:generate go run ../../../../../cmd/asyncapi-codegen -g user,types -p main -i ../../asyncapi.yaml -o ./user.gen.go
 
 package main
 
 import (
 	"context"
 
-	"github.com/lerenn/asyncapi-codegen/pkg/extensions/brokers/natsjetstream"
+	"github.com/lerenn/asyncapi-codegen/pkg/extensions/brokers/nats"
 	"github.com/lerenn/asyncapi-codegen/pkg/extensions/loggers"
 	"github.com/lerenn/asyncapi-codegen/pkg/extensions/middlewares"
 )
@@ -13,11 +13,10 @@ import (
 func main() {
 	// Instantiate a NATS controller with a logger
 	logger := loggers.NewText()
-	broker, err := natsjetstream.NewController(
-		"nats://nats-jetstream:4222",       // Set URL to broker
-		natsjetstream.WithLogger(logger),   // Attach an internal logger
-		natsjetstream.WithStream("ping"),   // Set the stream used
-		natsjetstream.WithConsumer("ping"), // Create the corresponding consumer
+	broker, err := nats.NewController(
+		"nats://nats:4222",                // Set URL to broker
+		nats.WithLogger(logger),           // Attach an internal logger
+		nats.WithQueueGroup("ping-users"), // Set a specific queue group to avoid collisions
 	)
 	if err != nil {
 		panic(err)
