+++ examples/ping/v2/nats/user/user.gen.go
