+++ pkg/pipeline/examples.go
@@ -1,14 +1,9 @@
 package pipeline
 
 import (
-	"os"
 	"strings"
 
 	"dagger.io/dagger"
-)
-
-const (
-	examplesPath = "./examples/"
 )
 
 // Examples returns a container that runs all examples.
@@ -16,9 +11,9 @@ func Examples(client *dagger.Client, brokers map[string]*dagger.Service) map[str
 	containers := make(map[string]*dagger.Container, 0)
 
 	// Set examples
-	for _, p := range examplesPaths() {
+	for _, p := range directoriesAtSublevel(2, "./examples") {
 		// Get corresponding broker
-		brokerName := strings.Split(p, "/")[1]
+		brokerName := strings.Split(p, "/")[4]
 
 		// Set app container
 		app := client.Container().
@@ -28,7 +23,7 @@ func Examples(client *dagger.Client, brokers map[string]*dagger.Service) map[str
 			// Set broker as dependency
 			WithServiceBinding(brokerName, brokers[brokerName]).
 			// Execute command
-			WithExec([]string{"go", "run", examplesPath + p + "/app"}).
+			WithExec([]string{"go", "run", p + "/app"}).
 			// Add exposed port to let know when the service is ready
 			WithExposedPort(1234).
 			// Set as service
@@ -45,41 +40,11 @@ func Examples(client *dagger.Client, brokers map[string]*dagger.Service) map[str
 			// Add app as dependency of user
 			WithServiceBinding("app", app).
 			// Execute command
-			WithExec([]string{"go", "run", examplesPath + p + "/user"})
+			WithExec([]string{"go", "run", p + "/user"})
 
 		// Add user containers to containers
 		containers[p] = user
 	}
 
 	return containers
-}
-
-func examplesPaths() []string {
-	paths := make([]string, 0)
-
-	examples, err := os.ReadDir(examplesPath)
-	if err != nil {
-		panic(err)
-	}
-
-	for _, e := range examples {
-		if !e.Type().IsDir() {
-			continue
-		}
-
-		brokers, err := os.ReadDir("./examples/" + e.Name())
-		if err != nil {
-			panic(err)
-		}
-
-		for _, b := range brokers {
-			if !b.Type().IsDir() {
-				continue
-			}
-
-			paths = append(paths, e.Name()+"/"+b.Name())
-		}
-	}
-
-	return paths
 }
