+++ pkg/pipeline/tests.go
@@ -1,8 +1,6 @@
 package pipeline
 
 import (
-	"os"
-
 	"dagger.io/dagger"
 )
 
@@ -11,7 +9,7 @@ func Tests(client *dagger.Client, brokers map[string]*dagger.Service) map[string
 	containers := make(map[string]*dagger.Container, 0)
 
 	// Set examples
-	for _, p := range testsPaths(3, "./test") {
+	for _, p := range directoriesAtSublevel(3, "./test") {
 		t := client.Container().
 			// Add base image
 			From(GolangImage).
@@ -27,35 +25,4 @@ func Tests(client *dagger.Client, brokers map[string]*dagger.Service) map[string
 	}
 
 	return containers
-}
-
-func testsPaths(level int, path string) []string {
-	paths := make([]string, 0)
-
-	tests, err := os.ReadDir(path)
-	if err != nil {
-		panic(err)
-	}
-
-	if level == 0 {
-		for _, t := range tests {
-			if !t.Type().IsDir() {
-				continue
-			}
-
-			paths = append(paths, path+"/"+t.Name())
-		}
-
-		return paths
-	}
-
-	for _, t := range tests {
-		if !t.Type().IsDir() {
-			continue
-		}
-
-		paths = append(paths, testsPaths(level-1, path+"/"+t.Name())...)
-	}
-
-	return paths
 }
